CH-14 Refactor tests to be more useful as documentation
