+++ build.gradle
@@ -11,6 +11,7 @@ repositories {
 
 dependencies {
     testImplementation 'junit:junit:4.13'
+    testImplementation 'org.mockito:mockito-core:3.9.0'
 }
 
 // See https://github.com/JetBrains/gradle-intellij-plugin/
